+++ changelog.md
@@ -10,3 +10,4 @@
 - Fix items being lost when switching tab in the engineering station
 - Fix item duplication in the engineering station upgrade tab
 - Fix some space farm interactions
+- Fix machines sending fluid updates to every player on the server